+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## Added
+
+- Added support for Ubuntu 22.04.
+- Added support for Debian 11.
+- Added shellcheck disable comments.
+
+### Changed
+
+- ⚠️ Now installs Java 17 or Java 11 (dependent on distribution) instead of Java 13.
+- ⚠️ Removed support for Ubuntu 16.04.
+- ⚠️ Removed support of Linux Mint 18 due to EOL.
+- Installs redis-server instead of keydb.
+
 ## [1.0.0] 2022-03-07
 
 Initial working release.